Update Refactored Project.py
Added huge amounts of cosmetic features including a background, button textures, highlighting for cards and buttons.
I have also added the ability to play cards, end turn, buy units from the shop and close the game.
There is currently a bug if you play an ogre as the last card before the cap it overcomes the card cap